fix: correctly support `--snapshotId` flag (#866)
https://coveord.atlassian.net/browse/CDX-1078

+++ packages/cli/src/commands/org/resources/preview.ts
@@ -100,8 +100,13 @@ export default class Preview extends Command {
     return flags.previewLevel === PreviewLevelValue.Detailed;
   }
 
+  private async shouldDeleteSnapshot() {
+    return !(await this.parse(Preview)).flags.snapshotId;
+  }
   private async cleanup(snapshot: Snapshot, project: Project) {
-    await snapshot.delete();
+    if (await this.shouldDeleteSnapshot()) {
+      await snapshot.delete();
+    }
     project.deleteTemporaryZipFile();
   }
 
@@ -117,7 +122,7 @@ export default class Preview extends Command {
 
           Once the snapshot is created, you can preview it with the following command:
 
-            ${blueBright`coveo org:resources:preview -t ${target} -s ${snapshot.id}`}
+            ${blueBright`coveo org:resources:preview -o ${target} -s ${snapshot.id}`}
 
             `
       );
@@ -129,6 +134,7 @@ export default class Preview extends Command {
     return {
       deleteMissingResources: flags.showMissingResources,
       waitUntilDone: {wait: flags.wait},
+      ...(flags.snapshotId && {snapshotId: flags.snapshotId}),
     };
   }
